CRISPR/requirements.txt
biopython
pysam
h5py
numpy
regex
